+++ fiddles/jquery/fiddle-0069-BootstrapStickyNavBars/assets/js/components/column-slider/column-slider.js
@@ -1,0 +1,19 @@
+jQuery(function ($) {
+
+  'use strict';
+
+
+  $('.tr-sticky').theiaStickySidebar({
+    additionalMarginTop: 0
+  });
+
+  $('#column-slider').on('slide.bs.carousel', function(){
+    $('.tr-sticky').theiaStickySidebar().disable();
+    $('html, body').animate({ scrollTop: 500 }, 'slow');
+    window.setTimeout(function() {
+      $('.tr-sticky').theiaStickySidebar().enable();
+    }, 500);
+  });
+
+
+});