+++ fiddles/jquery/fiddle-0069-BootstrapStickyNavBars/assets/js/components/header/header.js
@@ -1,0 +1,17 @@
+jQuery(function ($) {
+
+  'use strict';
+
+  $('.navbar-area').magnificPopup({
+    type: 'inline',
+    midClick: true,
+    mainClass: 'mfp-fade',
+  });
+
+  $('.navbar-area-right').magnificPopup({
+    type: 'inline',
+    midClick: true,
+    mainClass: 'mfp-fade'
+  });
+
+});