+++ fiddles/jquery/fiddle-0069-BootstrapStickyNavBars/assets/js/components/side-bar/side-bar.js
@@ -1,0 +1,14 @@
+
+jQuery(function ($) {
+
+  'use strict';
+
+  $('.tr-hamburger').magnificPopup({
+    mainClass: 'asd-menu mfp-slide-left',
+    type:'inline',
+    midClick: true,
+    closeMarkup: '<span class="aside-menu-close-ic mfp-close"></span>',
+    removalDelay: 300
+  });
+
+});